src: split lightdb into state and stream
The monolith lightdb module is split into lightdb_state,
lightdb_stream, and payload_utils.

Additionaly, the lightdb_*_internal() helper functions were
removed. These functions only abstracted 1 or 2 lines of common
code each, but took 8-10 arguments, meaning that removing
the helper functions actually improves readability by reducing
the total lines of code by about 100.
