+++ include/golioth.h
@@ -12,9 +12,11 @@ extern "C" {
 #include "golioth_status.h"
 #include "golioth_client.h"
 #include "golioth_log.h"
-#include "golioth_lightdb.h"
+#include "golioth_lightdb_state.h"
+#include "golioth_lightdb_stream.h"
 #include "golioth_rpc.h"
 #include "golioth_ota.h"
+#include "golioth_payload_utils.h"
 #include "golioth_fw_update.h"
 #include "golioth_settings.h"
 #include "golioth_debug.h"
