+++ port/esp_idf/components/golioth_sdk/CMakeLists.txt
@@ -34,9 +34,11 @@ idf_component_register(
         "${sdk_src}/golioth_status.c"
         "${sdk_src}/golioth_coap_client.c"
         "${sdk_src}/golioth_log.c"
-        "${sdk_src}/golioth_lightdb.c"
+        "${sdk_src}/golioth_lightdb_state.c"
+        "${sdk_src}/golioth_lightdb_stream.c"
         "${sdk_src}/golioth_rpc.c"
         "${sdk_src}/golioth_ota.c"
+        "${sdk_src}/golioth_payload_utils.c"
         "${sdk_src}/golioth_fw_update.c"
         "${sdk_src}/golioth_statistics.c"
         "${sdk_src}/golioth_settings.c"
