+++ port/linux/golioth_sdk/CMakeLists.txt
@@ -38,9 +38,11 @@ set(sdk_srcs
     "${sdk_src}/golioth_status.c"
     "${sdk_src}/golioth_coap_client.c"
     "${sdk_src}/golioth_log.c"
-    "${sdk_src}/golioth_lightdb.c"
+    "${sdk_src}/golioth_lightdb_state.c"
+    "${sdk_src}/golioth_lightdb_stream.c"
     "${sdk_src}/golioth_rpc.c"
     "${sdk_src}/golioth_ota.c"
+    "${sdk_src}/golioth_payload_utils.c"
     "${sdk_src}/golioth_fw_update.c"
     "${sdk_src}/golioth_statistics.c"
     "${sdk_src}/golioth_settings.c"
